backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
openpyxl>=3.1.0
pandas>=2.0.0
matplotlib>=3.7.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
contextily>=1.4.0
geopandas>=0.14.0
anthropic>=0.39.0
Pillow>=10.0.0
shapely>=2.0.0
python-pptx>=1.0.0
